requirements.txt
oandapyV20
pandas
psycopg2-binary==2.9.11
python-dateutil
pytz
requests
requests==2.32.3
python-dotenv
